More warnings removed
git-svn-id: file:///Volumes/GecodeGitMigration/gecode-svn-mirror/gecode/trunk@15120 e85b7adc-8362-4630-8c63-7469d557c915

+++ gecode/gist/spacenode.hpp
@@ -174,7 +174,7 @@ namespace Gecode { namespace Gist {
     SpaceNode* p = getParent(na);
     if (p == NULL)
       return -1;
-    for (int i=p->getNumberOfChildren(); i--;)
+    for (int i=static_cast<int>(p->getNumberOfChildren()); i--;)
       if (p->getChild(na,i) == this)
         return i;
     GECODE_NEVER;
